+++ source/HashList.cpp
@@ -7,18 +7,11 @@
 
 #include "../headers/HashList.hpp"
 
-HashList::HashList(int hash, int num_keys, KeyNode* keys) : hash(hash), num_keys(num_keys), keys(keys) {}
+HashList::HashList(int i_num_keys, KeyNode* i_keys)
+        : num_keys(i_num_keys), keys(i_keys) {}
 
 HashList::~HashList() {
 
-}
-
-int HashList::get_hash() const {
-    return hash;
-}
-
-void HashList::set_hash(int i_hash) {
-    HashList::hash = i_hash;
 }
 
 int HashList::get_num_keys() const {