+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+# Change Log
+
+## 2019-03-21
+
+#### v0.0.1
+
+###### Additions
+
+- Add comments and documentation to node, list,
+table,  output classes
+
+- Add TODOs marking where and what functions to implement
+
 ## 2019-03-20
 
+#### v0.0.1
+
+###### Additions
+
+- Add classes, class members, and associated gets/sets,
+constructor, and destructor
+
+- Add doc and testing files
+
 #### Initial Commit
+
+###### Additions
+
+- Add LICENSE and .gitignore